more code refactoring
- Move `DeadComponentError`, `Pointer` and `WeakPointer` to the module `pointer.rs`
(previously `weak.rs`)
- Update doc links for WeakPointer.
- Move `Item`, `Iter` & `IterMut` to the module `storage.rs`, since it was just used there.
- Made `PointerData` explicitly public to the crate (before it was implicit).
